+++ src/ui/components/ConsoleSurface.tsx
@@ -5,12 +5,13 @@ import { Transaction } from './Transaction';
 // import { Interpreter } from './InterpreterHacky';
 
 interface ConsoleSurfaceProps {
-	sendCurrentRequest: () => void;
+	sendRequest: (tId: number) => void;
 	setProtocol: (protocolId: string) => void;
-	updateUI: (vizItem: IVisualizationItem<any>, vizCurrent: IVisualization) => void;
-	openTextDocument: (docOptions: OpenTextDocumentOptions, viz: IVisualizationItem<BytesValue>) => void;
+	updateUI: (vizItem: IVisualizationItem<any>, tId: number) => void;
+	openTextDocument: (docOptions: OpenTextDocumentOptions, vizId: number, vizItem: IVisualizationItem<BytesValue>) => void;
 	getFunctionPreview: (command: string) => Promise<IVisualizationItem<any> | null>;
 	rerun: () => void;
+	clear: () => void;
 
 	currentRequest: IVisualization;
 	history: IVisualization[];
@@ -26,28 +27,36 @@ export class ConsoleSurface extends React.Component<ConsoleSurfaceProps> {
 
 				// TODO this logic is hardcoded and may be wrong. Currently, every keypress
 				// on an element other than a text area will trigger key bindings.
-				if (target.nodeName === 'TEXTAREA') {
+				if (target.nodeName === 'TEXTAREA' || target.nodeName === 'BUTTON') {
 					return;
 				}
 			}
 
 			if (e.key === 'Enter') {
-				this.props.sendCurrentRequest();
+				this.props.sendRequest(this.props.currentRequest.transaction.id!);
 			}
 		});
+	}
+
+	cbOpenTextDocument = (viz: IVisualization): ((docOptions: OpenTextDocumentOptions, vizItem: IVisualizationItem<BytesValue>) => void) => {
+		return (
+			docOptions: OpenTextDocumentOptions,
+			vizItem: IVisualizationItem<BytesValue>
+		) => this.props.openTextDocument(docOptions, viz.transaction.id!, vizItem);
 	}
 
 	render() {
 		// const history = this.props.history;
 		const {
 			history, currentRequest, rerunQueue, allProtocols,
-			sendCurrentRequest, setProtocol, updateUI, openTextDocument, getFunctionPreview
+			sendRequest, setProtocol, updateUI, getFunctionPreview,
+			rerun, clear,
 		} = this.props;
 
 		let reqCount = 0;
 		let resCount = 0;
 
-		let content = history.map( (item) => {
+		let content = history.map( (item, indexHistory) => {
 			let index;
 			if (item.context === 'outgoing') {
 				index = reqCount++;
@@ -58,15 +67,16 @@ export class ConsoleSurface extends React.Component<ConsoleSurfaceProps> {
 			// TODO: Some of the method props not needed in this case, remove
 			return <Transaction
 				index={index}
-				key={reqCount + resCount}
-				readOnly={true}
+				isCurrent={false}
+				key={indexHistory}
+				readOnly={item.context === 'incoming'}
 				visualization={item}
 				allProtocols={allProtocols}
 
-				updateUI={updateUI}
+				updateUI={(vizItem) => updateUI(vizItem, item.transaction.id!)}
 				setProtocol={setProtocol}
-				sendCurrentRequest={sendCurrentRequest}
-				openTextDocument={openTextDocument}
+				sendCurrentRequest={() => sendRequest(item.transaction.id!)}
+				openTextDocument={this.cbOpenTextDocument(item)}
 				getFunctionPreview={getFunctionPreview} />;
 		} );
 
@@ -78,28 +88,30 @@ export class ConsoleSurface extends React.Component<ConsoleSurfaceProps> {
 					index={index}
 					key={index}
 					readOnly={true}
+					isCurrent={false}
 					visualization={item}
 					allProtocols={allProtocols}
 
-					updateUI={updateUI}
+					updateUI={(vizItem) => null}
 					setProtocol={setProtocol}
-					sendCurrentRequest={sendCurrentRequest}
-					openTextDocument={openTextDocument}
+					sendCurrentRequest={() => null}
+					openTextDocument={this.cbOpenTextDocument(item)}
 					getFunctionPreview={getFunctionPreview} />;
 			});
 		} else {
 			content.push(
 				<Transaction
 					index={reqCount++}
-					key={reqCount + resCount}
+					key={history.length}
 					readOnly={false}
+					isCurrent={true}
 					visualization={currentRequest}
 					allProtocols={allProtocols}
 
-					updateUI={updateUI}
+					updateUI={(vizItem) => updateUI(vizItem, currentRequest.transaction.id!)}
 					setProtocol={setProtocol}
-					sendCurrentRequest={sendCurrentRequest}
-					openTextDocument={openTextDocument}
+					sendCurrentRequest={() => sendRequest(currentRequest.transaction.id!)}
+					openTextDocument={this.cbOpenTextDocument(currentRequest)}
 					getFunctionPreview={getFunctionPreview} />
 			);
 		}
@@ -107,7 +119,7 @@ export class ConsoleSurface extends React.Component<ConsoleSurfaceProps> {
 		return <div id="content">
 			<div style={({'marginTop':'15px'})}>🧪 Thanks for trying out the <i>alpha</i> version of VSConnect! All feedback is welcome on <a target="_blank" href="https://github.com/jcoc611/vsconnect/issues">Github</a>.</div>
 			{content}
-			<div style={({'opacity': '0.7'})}>{rerunContent}</div>
+			<div style={({'opacity': '0.6'})}>{rerunContent}</div>
 		</div>;
 	}
 }