+++ src/ui/console.tsx
@@ -17,15 +17,24 @@ state.getAllProtocols().then( (protocols) => {
 				rerunQueue={ state.getRerunQueue() }
 				allProtocols={protocols}
 				currentRequest={state.getCurrentRequest()}
-				sendCurrentRequest={ state.sendCurrentRequest.bind(state) }
+				sendRequest={ state.sendRequest.bind(state) }
 				setProtocol={ state.setProtocol.bind(state) }
 				updateUI={ state.updateUI.bind(state) }
 				openTextDocument={ state.openTextDocument.bind(state) }
 				getFunctionPreview={ state.getFunctionPreview.bind(state) }
-				rerun={ state.rerun.bind(state) } />,
+				rerun={ state.rerun.bind(state) }
+				clear={ state.clear.bind(state) } />,
 			document.getElementById("content-wrapper")
 		);
 		window.scrollTo(0,document.body.scrollHeight);
+
+		ReactDOM.render(
+			<div className="topBar">
+				<button onClick={ state.rerun.bind(state) } title='Resend every request one after the other'>🔄 Rerun</button>
+				<button onClick={ state.clear.bind(state) } title='Delete everything in this tab'>🗑️ Clear</button>
+			</div>,
+			document.getElementById("topBar-wrapper")
+		);
 
 		let contextMenuWrapper = document.getElementById('contextmenu-wrapper');
 		ReactDOM.render(